Address remaining review comments
- Fix getOptionalTensorDescriptor to throw on null elementCount when
  descSource is null, consistent with all other getters
- Rename 244-char test to GetTensorDescriptorsReturnsExpectedOrder
- Remove unnecessary TearDown() from test fixture (unique_ptr resets
  are automatic at fixture destruction)
- Simplify ComputeDataTypePreserved graph test to use required-only
  helper instead of creating all 28 tensors
- Fix stats_tensor_uid comparison to use has_value()/value() pattern
  for clearer failure diagnostics
- Add stats tensor shape [2,4,128,1] and data_type verification to
  SdpaFpropWithStatsRoundTrip integration test
- Remove redundant static_cast<int32_t> in packer (value() already
  returns int32_t)

+++ projects/hipdnn/frontend/include/hipdnn_frontend/detail/SdpaFpropPacker.hpp
@@ -252,7 +252,7 @@ inline Error createSdpaFpropOperation(
     }
     if(attributes.max_seq_len_kv.has_value())
     {
-        auto val = static_cast<int32_t>(attributes.max_seq_len_kv.value());
+        auto val = attributes.max_seq_len_kv.value();
         HIPDNN_CHECK_ERROR(setDescriptorAttrScalar(opDesc.get(),
                                                    HIPDNN_ATTR_SDPA_FPROP_MAX_SEQ_LEN_KV_EXT,
                                                    HIPDNN_TYPE_INT32,
